Addressed comments and updated parent readme
