+++ machine-learning-notebooks/transfer-learning-custom-azureml/camera/__init__.py
@@ -1,1 +1,1 @@
-from .camera import Camera
+from .capture_camera import Camera
